--- a/ChangeLog
+++ b/ChangeLog
@@ -234,7 +234,7 @@ Release date: TBA
 * Allow a `.` as a prefix for Sphinx name resolution.
 
 * Checkers must now keep a 1 to 1 relationship between "msgid" (ie: C1234) and "symbol" (ie : human-readable-symbol)
-* In checkers, an old_names can now be used for multiple new messages
+* In checkers, an old_names can now be used for multiple new messages and pylint is now a little faster
 
 Caused by #1164. It means if you do a partial old_names for a message definition an exception will tell you that you
 must rename the associated identification.