--- a/tests/message/unittest_message_id_store.py
+++ b/tests/message/unittest_message_id_store.py
@@ -52,6 +52,26 @@ def test_register_message_definitions(empty_msgid_store, message_definitions):
     assert len(empty_msgid_store) == number_of_msgid
 
 
+def test_add_msgid_and_symbol(empty_msgid_store):
+    empty_msgid_store.add_msgid_and_symbol("E1235", "new-sckiil")
+    empty_msgid_store.add_legacy_msgid_and_symbol("C1235", "old-sckiil", "E1235")
+    assert len(empty_msgid_store) == 2
+    message_ids = empty_msgid_store.get_active_msgids("E1235")
+    assert len(message_ids) == 1
+    assert message_ids[0] == "E1235"
+    message_ids = empty_msgid_store.get_active_msgids("old-sckiil")
+    assert len(message_ids) == 1
+    assert message_ids[0] == "E1235"
+    assert empty_msgid_store.get_symbol("C1235") == "old-sckiil"
+    assert empty_msgid_store.get_symbol("E1235") == "new-sckiil"
+    assert empty_msgid_store.get_msgid("old-sckiil") == "C1235"
+    assert empty_msgid_store.get_msgid("new-sckiil") == "E1235"
+    with pytest.raises(KeyError) as e:
+        empty_msgid_store.get_symbol("C1234")
+    with pytest.raises(KeyError) as e:
+        empty_msgid_store.get_msgid("not-exist")
+
+
 def test_duplicate_symbol(empty_msgid_store):
     empty_msgid_store.add_msgid_and_symbol("W1234", "warning-symbol")
     with pytest.raises(InvalidMessageError) as error: