#  Description

Permit to have multiple messages with the same old name. There are optimisations done that make Pylint a little faster.

## Type of Changes
|   | Type |
| ------------- | ------------- |
| ✓  | :sparkles: New feature |

## Related Issue

This is it, after this PR the real review for #1164 can start.