@Pierre-Sassoulas 2019-08-19
@PCManticore this one is ready for review, it should be a piece of cake after #2992 ;)